// src/app/core/models/item.ts
export class Item {
  id: number;
  code: number;
  name: string;
  brandName: string;
  description: string;
  price: string;
  displayPrice: string;
  hasVat: any;
  isSenior: any;
  weighted: any;
  packaging: string;
  packageMeasurement: number;
  sizing: string;
  packageMinimum: number;
  packageIntervals: number;
  availableOn: number;
  imageKey: string;
  slug: string;
  enabled: any;
  partner_id: number;
  dateCreated: number;
  dateUpdated: number;
  category1: number;
  category2: number;
  category3: number;
  quantity: number;
}